… CHANGELOG.md
… instrumentation/github.com/emicklei/go-restful/otelrestful/internal/semconv/env_test.go
… instrumentation/github.com/emicklei/go-restful/otelrestful/internal/semconvutil/httpconv_test.go
… instrumentation/github.com/gin-gonic/gin/otelgin/internal/semconv/env_test.go
… instrumentation/github.com/gin-gonic/gin/otelgin/internal/semconvutil/httpconv_test.go
… instrumentation/github.com/gorilla/mux/otelmux/internal/semconv/env_test.go
… instrumentation/github.com/gorilla/mux/otelmux/internal/semconv/v1.20.0.go
… instrumentation/github.com/gorilla/mux/otelmux/internal/semconvutil/httpconv.go
… instrumentation/github.com/gorilla/mux/otelmux/internal/semconvutil/httpconv_test.go
… instrumentation/github.com/labstack/echo/otelecho/internal/semconvutil/httpconv.go
… instrumentation/github.com/labstack/echo/otelecho/internal/semconvutil/httpconv_test.go
… instrumentation/net/http/httptrace/otelhttptrace/internal/semconv/env.go
… instrumentation/net/http/httptrace/otelhttptrace/internal/semconv/env_test.go
… instrumentation/net/http/httptrace/otelhttptrace/internal/semconv/v1.20.0.go
… instrumentation/net/http/httptrace/otelhttptrace/internal/semconvutil/httpconv.go
… instrumentation/net/http/httptrace/otelhttptrace/internal/semconvutil/httpconv_test.go
… instrumentation/net/http/otelhttp/internal/semconv/env.go
… instrumentation/net/http/otelhttp/internal/semconv/env_test.go
… instrumentation/net/http/otelhttp/internal/semconv/v1.20.0.go
… instrumentation/net/http/otelhttp/internal/semconvutil/httpconv.go
… instrumentation/net/http/otelhttp/internal/semconvutil/httpconv_test.go
… internal/shared/semconv/env.go.tmpl
… internal/shared/semconv/env_test.go.tmpl
… internal/shared/semconv/v1.20.0.go.tmpl
… internal/shared/semconvutil/httpconv.go.tmpl
… internal/shared/semconvutil/httpconv_test.go.tmpl