Good day,

While turning using `OTEL_SEMCONV_STABILITY_OPT_IN=http/dup` we noticed an increase of object allocations.

This PR proposes a change where instead of changing the (for example) `OldHTTPServer.RequestTraceAttrs` and `CurrentHTTPServer.RequestTraceAttrs` allocate we pass the slice created by `OldHTTPServer.RequestTraceAttrs` as an argument to `CurrentHTTPServer.RequestTraceAttrs`. This allows us to reduce the amount of allocation a little bit (see below).

Let me know what you think and thanks for reviewing this PR 😄

benchstat from `instrumentation/net/http/otelhttp/internal/semconv` with env var `OTEL_SEMCONV_STABILITY_OPT_IN=http/dup`.
```goos: linux
goarch: amd64
pkg: go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp/internal/semconv
cpu: AMD Ryzen 7 Pro 7735U with Radeon Graphics     
                            │ a50074d5f20689a804d628d482d0b0cdfe73c623-dup.txt │             new-dup.txt             │
                            │                      sec/op                      │   sec/op     vs base                │
HTTPServerRequest-16                                              1046.5n ± 1%   932.8n ± 2%  -10.86% (p=0.001 n=10)
RecordMetrics/empty-16                                             7.202n ± 0%   7.262n ± 0%   +0.84% (p=0.000 n=10)
RecordMetrics/nil_meter-16                                         843.5n ± 1%   858.1n ± 1%   +1.72% (p=0.003 n=10)
RecordMetrics/with_Meter-16                                        842.7n ± 2%   867.6n ± 2%   +2.96% (p=0.000 n=10)
geomean                                                            270.5n        266.5n        -1.50%

                            │ a50074d5f20689a804d628d482d0b0cdfe73c623-dup.txt │              new-dup.txt               │
                            │                       B/op                       │     B/op      vs base                  │
HTTPServerRequest-16                                            2.938Ki ± 0%     2.062Ki ± 0%  -29.79% (p=0.000 n=10)
RecordMetrics/empty-16                                            0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
RecordMetrics/nil_meter-16                                      1.313Ki ± 0%     1.313Ki ± 0%        ~ (p=1.000 n=10) ¹
RecordMetrics/with_Meter-16                                     1.313Ki ± 0%     1.313Ki ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                                      ²                  -8.46%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ a50074d5f20689a804d628d482d0b0cdfe73c623-dup.txt │             new-dup.txt              │
                            │                    allocs/op                     │ allocs/op   vs base                  │
HTTPServerRequest-16                                              3.000 ± 0%     2.000 ± 0%  -33.33% (p=0.000 n=10)
RecordMetrics/empty-16                                            0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
RecordMetrics/nil_meter-16                                        8.000 ± 0%     8.000 ± 0%        ~ (p=1.000 n=10) ¹
RecordMetrics/with_Meter-16                                       8.000 ± 0%     8.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                                      ²                -9.64%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```
Benchstats with `OTEL_SEMCONV_STABILITY_OPT_IN` unset:
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp/internal/semconv
cpu: AMD Ryzen 7 Pro 7735U with Radeon Graphics     
                            │ a50074d5f20689a804d628d482d0b0cdfe73c623.txt │              new.txt               │
                            │                    sec/op                    │   sec/op     vs base               │
HTTPServerRequest-16                                           442.6n ± 2%   444.0n ± 2%       ~ (p=0.669 n=10)
RecordMetrics/empty-16                                         7.229n ± 0%   7.261n ± 0%  +0.44% (p=0.000 n=10)
RecordMetrics/nil_meter-16                                     452.1n ± 1%   457.8n ± 1%  +1.27% (p=0.034 n=10)
RecordMetrics/with_Meter-16                                    453.2n ± 2%   450.7n ± 1%       ~ (p=0.615 n=10)
geomean                                                        160.0n        160.6n       +0.38%

                            │ a50074d5f20689a804d628d482d0b0cdfe73c623.txt │               new.txt               │
                            │                     B/op                     │    B/op     vs base                 │
HTTPServerRequest-16                                          704.0 ± 0%     704.0 ± 0%       ~ (p=1.000 n=10) ¹
RecordMetrics/empty-16                                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
RecordMetrics/nil_meter-16                                    672.0 ± 0%     672.0 ± 0%       ~ (p=1.000 n=10) ¹
RecordMetrics/with_Meter-16                                   672.0 ± 0%     672.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                  ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ a50074d5f20689a804d628d482d0b0cdfe73c623.txt │               new.txt               │
                            │                  allocs/op                   │ allocs/op   vs base                 │
HTTPServerRequest-16                                          1.000 ± 0%     1.000 ± 0%       ~ (p=1.000 n=10) ¹
RecordMetrics/empty-16                                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
RecordMetrics/nil_meter-16                                    4.000 ± 0%     4.000 ± 0%       ~ (p=1.000 n=10) ¹
RecordMetrics/with_Meter-16                                   4.000 ± 0%     4.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                  ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```