@dmathieu 2025-04-08
How about rebasing your PR against https://github.com/open-telemetry/opentelemetry-go-contrib/pull/6899 ?