@MrAlias 2025-04-08
> How about rebasing your PR against #6899 ?

Merged. Rebasing against `main` is needed.